cert_automation/requirements.txt
requests
python-dotenv
cryptography
pyOpenSSL
dnspython
PyYAML
paramiko